Move Hostnames to aikido_zen/storage
